Spostati i file del database
+ Modificato il nome di una task VSCode
+ Spostati i file di schema e popolazione db
    + Modificato il README per riflettere la modifica
    + Modificato il main.c per riflettere la modifica
+ Corretto warning in test_json-parser.c

NOTA! Aggiunta temporaneamente una directory src_backup al .gitignore

+++ source-code/backend/src/json-parser/test_json-parser.c
@@ -59,7 +59,7 @@ void json_c() {
     test_array_length = json_object_array_length(test_array);
     printf("Found %zu items\n", test_array_length);
 
-    for (int i=0; i<test_array_length; i++) {
+    for (size_t i=0; i<test_array_length; i++) {
         test_array_item = json_object_array_get_idx(test_array, i);
         printf ("%zu: %s\n", i+1, json_object_get_string(test_array_item));
     }
